+++ lib/core/gka.js
@@ -6,7 +6,7 @@
  */
 
 var path      = require("path"),
-    imagex    = require("./index"),
+    imagex    = require("../imagex"),
     util      = require("./util"), 
     writeFile = util.writeFile;
