+++ lib/imagex/crop.js
