+++ lib/imagex/file.js
