+++ lib/imagex/index.js
