+++ lib/imagex/info.js
