+++ lib/imagex/mini.js
